if servive revesion changed, mark service as updatable
