+++ pkg/microservice/aslan/core/environment/handler/service.go
@@ -22,6 +22,7 @@ import (
 
 	"github.com/gin-gonic/gin"
 
+	commonservice "github.com/koderover/zadig/pkg/microservice/aslan/core/common/service"
 	"github.com/koderover/zadig/pkg/microservice/aslan/core/environment/service"
 	"github.com/koderover/zadig/pkg/setting"
 	internalhandler "github.com/koderover/zadig/pkg/shared/handler"
@@ -34,7 +35,7 @@ func ListSvcsInEnv(c *gin.Context) {
 	envName := c.Param("name")
 	productName := c.Query("projectName")
 
-	ctx.Resp, ctx.Err = service.ListServicesInEnv(envName, productName, ctx.Logger)
+	ctx.Resp, ctx.Err = commonservice.ListServicesInEnv(envName, productName, ctx.Logger)
 }
 
 func ListSvcsInProductionEnv(c *gin.Context) {
@@ -43,7 +44,7 @@ func ListSvcsInProductionEnv(c *gin.Context) {
 	envName := c.Param("name")
 	productName := c.Query("projectName")
 
-	ctx.Resp, ctx.Err = service.ListServicesInProductionEnv(envName, productName, ctx.Logger)
+	ctx.Resp, ctx.Err = commonservice.ListServicesInProductionEnv(envName, productName, ctx.Logger)
 }
 
 func GetService(c *gin.Context) {
